requirements.txt
amqp==5.0.6
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==3.3.1
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-environ==0.5.0
django-extensions==3.1.3
django-filter==2.4.0
django-js-asset==1.2.2
django-livereload-server==0.3.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
entrypoints==0.3
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.8.6
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
razorpay==1.2.0
redis==3.5.3
requests==2.26.0
Send2Trash==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
